created per workday views base

+++ FarmManager.Model/Model/WorkDayCollecting.cs
@@ -6,7 +6,7 @@ public class WorkdayCollecting : BaseEntity
 {
     public int EmployeeId { get; set; }
     public Employee Employee { get; set; } = null!;
-    public int Quantity { get; set; }
+    public double Quantity { get; set; }
     public double Price { get; set; }
     public double RemainingToPay { get; set; }
     public int WorkdayId { get; set; }
